backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
celery==5.3.6
redis==5.0.1
kombu==5.3.4
flower==2.0.1
boto3==1.34.34
google-api-python-client==2.118.0
google-auth==2.28.1
google-generativeai==0.7.2
openai==1.12.0
yt-dlp>=2024.03.10
moviepy==1.0.3
Pillow==10.2.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
httpx==0.26.0
aiofiles==23.2.1
python-dotenv==1.0.1
structlog==24.1.0
sentry-sdk==1.40.6
prometheus-client==0.19.0
tenacity==8.2.3
schedule==1.2.1
APScheduler==3.10.4
aioredis==2.0.1
slowapi==0.1.9
